tcpv4tracer: get the kernel TGID instead of the PID
bfp_get_current_pid_tgid() returns a u64 containing
"current->tgid << 32 | current->pid".

We were storing the return value in a u32, which means we got
"current->pid". In kernel terms, the PID is actually what userspace
calls a thread ID.

What we actually want is what the userspace calls PID, and that's the
kernel's TGID.

Store the return value of bfp_get_current_pid_tgid() in a u64 and store
it right-shifted 32 bits so we get the actual PID (TGID).